backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-cloud-pubsub==2.21.5
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.117.0
google-generativeai==0.8.5
gunicorn==23.0.0
PyMuPDF==1.24.9
